Created functions,js file to import functions for displaying list
