+++ src/index.html
@@ -29,6 +29,7 @@
               placeholder="Add To Your List..."
               type="text"
               class="full-width wrapper input"
+              
             />
             <i class="fa fa-reply wrapper"></i>
           </form>
